+++ crates/cubecl-linalg/src/matmul/cmma/compute_loop/accumulators_first.rs
@@ -2,11 +2,12 @@ use cubecl_core as cubecl;
 use cubecl_core::prelude::*;
 
 use crate::matmul::cmma::{
-    base::{Fragments, Ids, SharedMemories},
+    base::{Fragments, SharedMemories},
     compute_loop::base::{
         get_smem_position_lhs, get_smem_position_rhs, load_tile_into_fragment, ComputeLoop,
     },
     config::ComptimeCmmaInfo,
+    ids::Ids,
 };
 
 pub(crate) struct AllAccumulatorsFirstComputeLoop {}
