+++ crates/cubecl-linalg/src/matmul/cmma/mod.rs
@@ -4,6 +4,7 @@ mod block_loop;
 pub(crate) mod compute_loop;
 pub(crate) mod config;
 pub(crate) mod cube_dispatch;
+mod ids;
 mod launch;
 pub(crate) mod load_shared_memory;
 pub(crate) mod write_output;
